refactor: changed ft_lstsize to size_t return

+++ src/ft_lstsize.c
@@ -1,20 +1,20 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_lstsize_bonus.c                                 :+:      :+:    :+:   */
+/*   ft_lstsize.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: troberts <troberts@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/19 03:55:02 by troberts          #+#    #+#             */
-/*   Updated: 2022/04/21 22:50:28 by troberts         ###   ########.fr       */
+/*   Updated: 2022/07/12 17:03:08 by troberts         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_lstsize(t_list *lst)
+size_t	ft_lstsize(t_list *lst)
 {
-	int	cnt;
+	size_t	cnt;
 
 	if (lst == NULL)
 		return (0);
